requirements.txt
pystray>=0.19.5
Pillow>=10.0.0
hidapi>=0.14.0
rivalcfg>=4.0.0
